[refactor]: feed -> survival 로직 분리 (#484)
