+++ 14th-team5-iOS/App/Sources/Presentation/Home/ViewControllers/MissionViewController.swift
@@ -1,0 +1,19 @@
+//
+//  MissionViewController.swift
+//  App
+//
+//  Created by 마경미 on 21.04.24.
+//
+
+import UIKit
+
+final class MissionViewController: UIViewController {
+    private let missionCollectionView: UICollectionView = UICollectionView(frame: .zero, collectionViewLayout: .init())
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        
+    }
+
+}
