+++ 14th-team5-iOS/App/Sources/Presentation/Home/ViewControllers/SurvivalViewController.swift
@@ -1,0 +1,153 @@
+//
+//  SurvivalViewController.swift
+//  App
+//
+//  Created by 마경미 on 21.04.24.
+//
+
+import UIKit
+
+import Core
+
+import RxSwift
+import RxDataSources
+
+final class SurvivalViewController: BaseViewController<SurvivalViewReactor>, UICollectionViewDelegate {
+    private let postCollectionView: UICollectionView = UICollectionView(frame: .zero, collectionViewLayout: UICollectionViewFlowLayout())
+    
+    private let refreshControl: UIRefreshControl = UIRefreshControl()
+    private let noPostView: NoPostTodayView = NoPostTodayView()
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        
+        navigationBarView.isHidden = true
+    }
+    
+    override func bind(reactor: SurvivalViewReactor) {
+        super.bind(reactor: reactor)
+        
+        bindInput(reactor: reactor)
+        bindOutput(reactor: reactor)
+    }
+    
+    override func setupUI() {
+        super.setupUI()
+        
+        view.addSubviews(postCollectionView, noPostView)
+    }
+    
+    override func setupAutoLayout() {
+        super.setupAutoLayout()
+        
+        postCollectionView.snp.makeConstraints {
+            $0.horizontalEdges.equalTo(view.safeAreaLayoutGuide)
+            $0.verticalEdges.equalToSuperview()
+        }
+        
+        noPostView.snp.makeConstraints {
+            $0.edges.equalToSuperview()
+        }
+    }
+    
+    override func setupAttributes() {
+        super.setupAttributes()
+        
+        postCollectionView.do {
+            $0.collectionViewLayout = createPostLayout()
+            $0.showsVerticalScrollIndicator = false
+            $0.backgroundColor = .clear
+            $0.refreshControl = refreshControl
+            $0.refreshControl?.tintColor = UIColor.bibbiWhite
+            $0.register(FeedCollectionViewCell.self, forCellWithReuseIdentifier: FeedCollectionViewCell.id)
+        }
+    }
+}
+
+extension SurvivalViewController {
+    private func bindInput(reactor: SurvivalViewReactor) {
+        postCollectionView.rx.setDelegate(self)
+            .disposed(by: disposeBag)
+        
+        Observable.just(())
+            .map { Reactor.Action.fetchPost }
+            .bind(to: reactor.action)
+            .disposed(by: disposeBag)
+
+        postCollectionView.rx.itemSelected
+            .throttle(RxConst.throttleInterval, scheduler: Schedulers.main)
+            .withUnretained(self)
+            .bind(onNext: {
+                $0.0.navigationController?.pushViewController(
+                    PostListsDIContainer().makeViewController(
+                        postLists: reactor.currentState.postSection,
+                        selectedIndex: $0.1
+                    ), animated: true)
+            })
+            .disposed(by: disposeBag)
+        
+        refreshControl.rx.controlEvent(.valueChanged)
+            .map { Reactor.Action.refresh }
+            .bind(to: reactor.action)
+            .disposed(by: disposeBag)
+    }
+    
+    private func bindOutput(reactor: SurvivalViewReactor) {
+        reactor.pulse(\.$postSection)
+            .observe(on: MainScheduler.instance)
+            .map(Array.init(with:))
+            .bind(to: postCollectionView.rx.items(dataSource: createPostDataSource()))
+            .disposed(by: disposeBag)
+        
+        reactor.pulse(\.$isRefreshEnd)
+            .withUnretained(self)
+            .observe(on: MainScheduler.instance)
+            .withUnretained(self)
+            .bind(onNext: { owner, _ in
+                DispatchQueue.main.asyncAfter(deadline: .now() + 0.5) {
+                    owner.refreshControl.endRefreshing()
+                }
+            })
+            .disposed(by: disposeBag)
+        
+        reactor.state.map { $0.isShowingNoPostTodayView }
+            .observe(on: MainScheduler.instance)
+            .distinctUntilChanged()
+            .map { !$0 }
+            .bind(to: noPostView.rx.isHidden)
+            .disposed(by: disposeBag)
+//
+//        reactor.pulse(\.$postSection)
+//            .withUnretained(self)
+//            .bind(onNext: {
+//                $0.0.timerView.reactor = TimerDIContainer().makeReactor(isSelfUploaded: reactor.currentState.isSelfUploaded, isAllUploaded: reactor.currentState.isAllFamilyMembersUploaded)
+//            })
+//            .disposed(by: disposeBag)
+    }
+}
+
+extension SurvivalViewController {
+    private func createPostLayout() -> UICollectionViewFlowLayout {
+        let layout = UICollectionViewFlowLayout()
+        layout.sectionInset = .zero
+        layout.itemSize = .init(width: UIScreen.main.bounds.width / 2 - 3, height: UIScreen.main.bounds.width / 2 - 3 + 36)
+        layout.minimumLineSpacing = HomeAutoLayout.FeedCollectionView.minimumLineSpacing
+        layout.minimumInteritemSpacing = HomeAutoLayout.FeedCollectionView.minimumInteritemSpacing
+        return layout
+    }
+    
+    private func createPostDataSource() -> RxCollectionViewSectionedReloadDataSource<PostSection.Model> {
+        return RxCollectionViewSectionedReloadDataSource<PostSection.Model>(
+            configureCell: { (_, collectionView, indexPath, item) in
+                switch item {
+                case .main(let data):
+                    guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: FeedCollectionViewCell.id, for: indexPath) as? FeedCollectionViewCell else {
+                        return UICollectionViewCell()
+                    }
+                    cell.reactor = SurvivalCellReactor(initialState: .init(postListData: data))
+                    return cell
+                }
+            })
+        
+    }
+}
